Move mock tests back to where they were before
