+++ tests/common/mod.rs
@@ -1,0 +1,3 @@
+pub use mock::MockCloudWatchClient;
+
+mod mock;
